+++ src/commands/misc/ping.ts
@@ -1,8 +1,11 @@
-module.exports = {
+import { Client, ChatInputCommandInteraction } from "discord.js";
+
+export default {
   name: 'ping',
   description: 'Replies with the bot ping!',
+  options: [], // Ensure options are defined, even if empty
 
-  callback: async (client, interaction) => {
+  callback: async (client: Client, interaction: ChatInputCommandInteraction) => {
     await interaction.deferReply();
 
     const reply = await interaction.fetchReply();
